improve efficiency of available version query
The availableVersions method obtains enough information to be able to present
retraction information for all versions since the current version. Because
retractions refer to versions at or before the retract directive we do not
need to get version information prior to the current version.

From the goproxy spec:

The new version must be higher than other release or pre-release versions; that
is, the @latest version query should resolve to the new version before retractions
are considered.